+++ Tests/NanoRoute.Tests/RouterBuilderTests.cs
@@ -26,7 +26,7 @@ namespace NanoRoute.Tests
 
         private RouterBuilder<TestRouter, RouterConfig> _routerBuilder = null!;
 
-        private Mock<Func<RouterBuilder<TestRouter, RouterConfig>, TestRouter>> _mockRouterFactory = null!;
+        private Mock<RouterFactoryDelegate<TestRouter, RouterConfig>> _mockRouterFactory = null!;
 
         private sealed class TestJsonPayload
         {
@@ -36,7 +36,7 @@ namespace NanoRoute.Tests
         [SetUp]
         public void Setup()
         {
-            _mockRouterFactory = new Mock<Func<RouterBuilder<TestRouter, RouterConfig>, TestRouter>>(MockBehavior.Strict);
+            _mockRouterFactory = new Mock<RouterFactoryDelegate<TestRouter, RouterConfig>>(MockBehavior.Strict);
             _routerBuilder = new RouterBuilder<TestRouter, RouterConfig>(_mockRouterFactory.Object);
 
             _mockRouterFactory